refactor: inline getDotAstroTypeRefs
