+++ packages/astro/src/content/index.ts
@@ -1,12 +1,7 @@
 export { CONTENT_FLAG, PROPAGATED_ASSET_FLAG } from './consts.js';
 export { attachContentServerListeners } from './server-listeners.js';
 export { createContentTypesGenerator } from './types-generator.js';
-export {
-	contentObservable,
-	getContentPaths,
-	getDotAstroTypeReferences,
-	hasAssetPropagationFlag,
-} from './utils.js';
+export { contentObservable, getContentPaths, hasAssetPropagationFlag } from './utils.js';
 export { astroContentAssetPropagationPlugin } from './vite-plugin-content-assets.js';
 export { astroContentImportPlugin } from './vite-plugin-content-imports.js';
 export { astroContentVirtualModPlugin } from './vite-plugin-content-virtual-mod.js';
