+++ packages/astro/src/content/utils.ts
@@ -14,9 +14,8 @@ import type {
 } from '../@types/astro.js';
 import { AstroError, AstroErrorData, MarkdownError, errorMap } from '../core/errors/index.js';
 import { isYAMLException } from '../core/errors/utils.js';
-import { CONTENT_FLAGS, CONTENT_TYPES_FILE, PROPAGATED_ASSET_FLAG } from './consts.js';
+import { CONTENT_FLAGS, PROPAGATED_ASSET_FLAG } from './consts.js';
 import { createImage } from './runtime-assets.js';
-import { ACTIONS_TYPES_FILE } from '../actions/consts.js';
 
 /**
  * Amap from a collection + slug to the local file path.
@@ -37,22 +36,6 @@ export const collectionConfigParser = z.union([
 		schema: z.any().optional(),
 	}),
 ]);
-
-export function getDotAstroTypeReferences({ fs, root, srcDir }: { fs: typeof fsMod, root: URL; srcDir: URL }) {
-	const { cacheDir } = getContentPaths({ root, srcDir });
-	let referenceDefs: string[] = [];
-	const typesFiles = [CONTENT_TYPES_FILE, ACTIONS_TYPES_FILE];
-	for (const typesFile of typesFiles) {
-		const url = new URL(typesFile, cacheDir); 
-		if (!fs.existsSync(url)) continue;
-		const typesRelativeToSrcDir = normalizePath(
-			path.relative(fileURLToPath(srcDir), fileURLToPath(url))
-		);
-		referenceDefs.push(`/// <reference path=${JSON.stringify(typesRelativeToSrcDir)} />`);
-	}
-
-	return referenceDefs;
-}
 
 export const contentConfigParser = z.object({
 	collections: z.record(collectionConfigParser),
